CalendarStore finalizado

+++ src/hooks/useCalendarStore.js
@@ -2,33 +2,72 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   onAddNewEvent,
   onDeleteEvent,
+  onLoadEvents,
   onSetActiveEvent,
   onUpdateEvent,
 } from '../store/calendar/calendarSlice.js';
+import { calendarApi } from '../apis';
+import { convertEventsToDateEvents } from '../helpers';
+import Swal from 'sweetalert2';
 
 export const useCalendarStore = () => {
   const dispatch = useDispatch();
   const { events, activeEvent } = useSelector((state) => state.calendar);
+  const { user } = useSelector((state) => state.auth);
 
   const setActiveEvent = (calendarEvent) => {
     dispatch(onSetActiveEvent(calendarEvent));
   };
 
   const startSavingEvent = async (calendarEvent) => {
-    // TODO: Llegar al backend
-
-    // TODO: todo ok
-
-    if (calendarEvent._id) {
-      dispatch(onUpdateEvent({ ...calendarEvent }));
-    } else {
-      dispatch(onAddNewEvent({ ...calendarEvent, _id: new Date().getTime() }));
+    try {
+      if (calendarEvent.id) {
+        await calendarApi.put(`/events/${calendarEvent.id}`, calendarEvent);
+        dispatch(onUpdateEvent({ ...calendarEvent, user }));
+        return;
+      }
+      const { data } = await calendarApi.post('/events', calendarEvent);
+      console.log(data);
+      dispatch(onAddNewEvent({ ...calendarEvent, id: data.evento.id, user }));
+    } catch (error) {
+      console.log(error);
+      Swal.fire('Error al guardar', error.response.data?.msg, 'error');
     }
   };
 
   const startDeleteEvent = async () => {
-    // TODO: llegar al backend
-    dispatch(onDeleteEvent());
+    try {
+      const { isConfirmed } = await Swal.fire({
+        title: 'Está seguro de que desea eliminar este evento?',
+        showDenyButton: true,
+
+        confirmButtonText: 'Eliminar',
+        denyButtonText: `Cancelar`,
+      });
+
+      if (!isConfirmed) throw new Error('Cancelaste la accion');
+
+      await calendarApi.delete(`/events/${activeEvent.id}`);
+      dispatch(onDeleteEvent());
+    } catch (error) {
+      console.log(error);
+      Swal.fire(
+        'Error al eliminar',
+        error.response?.data?.msg || error.message,
+        'error'
+      );
+    }
+  };
+
+  const startLoadingEvents = async () => {
+    try {
+      const { data } = await calendarApi.get('/events');
+      const events = convertEventsToDateEvents(data.eventos);
+      dispatch(onLoadEvents(events));
+    } catch (error) {
+      console.log('error cargando eventos');
+      console.log(error);
+    }
   };
   return {
     //Properties
@@ -39,5 +78,6 @@ export const useCalendarStore = () => {
     setActiveEvent,
     startSavingEvent,
     startDeleteEvent,
+    startLoadingEvents,
   };
 };
